Added registrtation, locked functions to non-users
